% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptestR-package.R
\docType{package}
\name{ptestR-package}
\alias{ptestR}
\alias{ptestR-package}
\title{ptestR: Permutation-Based Significance Testing for Regression Models}
\description{
ptestR wraps \code{\link[stats:glm]{stats::glm()}}, \code{\link[lme4:lmer]{lme4::lmer()}}, and binomial \code{\link[stats:glm]{stats::glm()}}
with a permutation loop to produce nonparametric p-values. For each model,
a null distribution of the test statistic is built by randomly rearranging
the outcome variable (\code{permNum} times); \code{p.perm} is the proportion of
permuted statistics at least as extreme as the observed one.

This approach requires far fewer distributional assumptions than standard
Wald or likelihood-ratio tests, making it well-suited to neuroimaging,
EEG, and other biomedical datasets with repeated measures and small samples.
\subsection{Main functions}{\tabular{lll}{
   Function \tab Model class \tab Test statistic \cr
   \code{\link[=grouped_perm_glm]{grouped_perm_glm()}} \tab Generalised linear model (\code{glm}) \tab t \cr
   \code{\link[=grouped_perm_glmm]{grouped_perm_glmm()}} \tab Linear mixed-effects model (\code{lmer}) \tab t \cr
   \code{\link[=grouped_perm_binoglm]{grouped_perm_binoglm()}} \tab Binomial logistic regression (\code{glm}) \tab z \cr
}


All three share the same call signature and return a tidy tibble.
}

\subsection{Typical grouped usage}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{df |>
  dplyr::group_by(State) |>
  tidyr::nest() |>
  dplyr::mutate(
    res = purrr::map(data, ~ grouped_perm_glmm(
      .x,
      formla      = feature ~ PMA + Sex + (1 | subject),
      var_to_perm = "feature",
      permNum     = 10000
    ))
  ) |>
  tidyr::unnest(res) |>
  dplyr::filter(term == "PMA")
}\if{html}{\out{</div>}}
}
}
\references{
França LGS, et al. (2024). Neonatal brain dynamic functional connectivity
in term and preterm infants and its association with early childhood
neurodevelopment. \emph{Nature Communications}, 15, 16.
\url{https://doi.org/10.1038/s41467-023-44050-z}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CoDe-Neuro/ptestR}
  \item Report bugs at \url{https://github.com/CoDe-Neuro/ptestR/issues}
}

}
\author{
\strong{Maintainer}: Lucas G. S. França \email{lucas.franca@kcl.ac.uk} (\href{https://orcid.org/0000-0003-0853-1319}{ORCID})

Authors:
\itemize{
  \item Yan Ge \email{yan.ge@kcl.ac.uk}
  \item Dafnis Batalle \email{dafnis.batalle@kcl.ac.uk} (\href{https://orcid.org/0000-0003-2097-979X}{ORCID})
}

}
\keyword{internal}
