% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_perm_binoglm.R
\name{grouped_perm_binoglm}
\alias{grouped_perm_binoglm}
\title{Permutation test for binomial logistic regression}
\usage{
grouped_perm_binoglm(tbl, formla, var_to_perm, permNum = 1000, seed = 42)
}
\arguments{
\item{tbl}{A data frame or a data frame extension (e.g. a tibble) containing
the variables in the model.}

\item{formla}{An object of class "formula" containing a symbolic description
of the logistic regression model to be fitted. see
\code{\link[stats]{formula}} for more details.}

\item{var_to_perm}{The columns of variables to permute.}

\item{permNum}{The number of permutations to generate.}

\item{seed}{A single value, interpreted as an integer for initializing a
random permutation generator. see \code{\link[base]{set.seed}} for more
details.}
}
\value{
It returns a tibble containing the following components:
\tabular{ll}{\code{term}\tab  The name of the regression term.\cr
\tab \cr
\code{estimate} \tab The estimated value of the regression term.\cr
\tab \cr
\code{statistic} \tab The value of a Z-statistic to use in a hypothesis
that the regression term is non-zero.\cr
\tab \cr
\code{p.value} \tab The two-sided p-value associated with the observed
statistic.\cr
\tab \cr
\code{p.perm} \tab The likelihood of observing the test statistic of the
original data among that of permutations.}
}
\description{
\code{grouped_perm_binoglm} is a nonparametric test of binomial
logistic regression models. It assesses the significance of coefficients by
permutation tests, which calculate the distribution of the test statistic by
randomly rearranging the observed data.
}
\examples{
counts <- c(10, 11, 8, 9, 6, 3, 5, 1)
gender <- c(0,0,0,0,0,1,1,1)
TBL <- data.frame(counts, gender)
grouped_perm_binoglm(TBL, gender ~ counts, "gender")
grouped_perm_binoglm(TBL, gender ~ counts, "gender", permNum = 500, seed = 1)

}
